# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClinicFlow AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics:
    get:
      operationId: listClinics
      tags: [clinics]
      summary: List all clinics
      responses:
        "200":
          description: List of clinics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clinic"
    post:
      operationId: createClinic
      tags: [clinics]
      summary: Register a new clinic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
  /clinics/{clinicId}:
    get:
      operationId: getClinic
      tags: [clinics]
      summary: Get clinic by ID
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
      responses:
        "200":
          description: Clinic details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
        "404":
          description: Not found
    put:
      operationId: updateClinic
      tags: [clinics]
      summary: Update clinic settings
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicUpdate"
      responses:
        "200":
          description: Updated clinic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"
...
  /clinics/{clinicId}/slots:
    get:
      operationId: listAvailableSlots
      tags: [clinics]
      summary: Get available slots for a date
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of slot strings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /clinics/{clinicId}/dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get admin dashboard stats for a clinic
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
components:
  parameters:
    ClinicIdParam:
      name: clinicId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Clinic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        address:
          type: string
        fee:
          type: string
        timings:
          type: string
        logoUrl:
          type: ["string", "null"]
        doctorSignatureUrl:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        doctorQualification:
          type: ["string", "null"]
        doctorSpecialization:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        clinicName:
          type: string
        contactEmail:
          type: string
        consultationFee:
          type: string
        operatingTimings:
          type: string
        clinicLogo:
          type: ["string", "null"]
        clinicPhoneNumber:
          type: ["string", "null"]
        workingSessions:
          type: array
          items:
            type: object
            properties:
              start:
                type: string
              end:
                type: string
            required: [start, end]
        createdAt:
          type: string
        workingDays:
          type: ["string", "null"]
        openingTime:
          type: ["string", "null"]
        closingTime:
          type: ["string", "null"]
        slotDuration:
          type: ["integer", "null"]
        googleConnected:
          type: boolean
        googleConnectedEmail:
          type: ["string", "null"]
        googleCalendarId:
          type: ["string", "null"]
        googleLastSyncAt:
          type: ["string", "null"]
      required: [id, name, email, address, fee, timings, createdAt, googleConnected]
    ClinicInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        password:
          type: string
          minLength: 6
        address:
          type: string
        fee:
          type: string
        timings:
          type: string
        logoUrl:
          type: ["string", "null"]
        doctorSignatureUrl:
          type: string
        doctorName:
          type: string
        doctorQualification:
          type: string
        doctorSpecialization:
          type: string
        phone:
          type: string
        clinicName:
          type: string
        contactEmail:
          type: string
        consultationFee:
          type: string
        operatingTimings:
          type: string
        clinicLogo:
          type: ["string", "null"]
        clinicPhoneNumber:
          type: ["string", "null"]
        workingSessions:
          type: array
          items:
            type: object
            properties:
              start:
                type: string
              end:
                type: string
            required: [start, end]
        workingDays:
          type: string
        openingTime:
          type: string
        closingTime:
          type: string
        slotDuration:
          type: integer
      required: [name, email, password, address, fee, timings]
    ClinicUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        address:
          type: string
        fee:
          type: string
        timings:
          type: string
        logoUrl:
          type: ["string", "null"]
        doctorSignatureUrl:
          type: string
        doctorName:
          type: string
        doctorQualification:
          type: string
        doctorSpecialization:
          type: string
        phone:
          type: string
        clinicName:
          type: string
        contactEmail:
          type: string
        consultationFee:
          type: string
        operatingTimings:
          type: string
        clinicLogo:
          type: ["string", "null"]
        clinicPhoneNumber:
          type: ["string", "null"]
        workingSessions:
          type: array
          items:
            type: object
            properties:
              start:
                type: string
              end:
                type: string
            required: [start, end]
        workingDays:
          type: string
        openingTime:
          type: string
        closingTime:
          type: string
        slotDuration:
          type: integer
    Appointment:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: string
        patientProblem:
          type: string
        appointmentDate:
          type: string
        sessionId:
          type: ["string", "null"]
        selectedTimeSlot:
          type: ["string", "null"]
        calendarEventId:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, pending_slot_selection, confirmed, booked, completed, cancelled]
        appointmentSource:
          type: string
          enum: [Online, Phone, Walk-in, Manual]
        patientAge:
          type: ["integer", "null"]
        patientGender:
          type: ["string", "null"]
        visitType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        doctorId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, clinicId, patientName, patientPhone, patientProblem, appointmentDate, status, createdAt]
...
    DashboardStats:
      type: object
      properties:
        clinic:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            planType:
              type: string
            subscriptionStatus:
              type: string
            expiryDate:
              type: ["string", "null"]
          required: [id, name, planType, subscriptionStatus]
        subscription:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            planType:
              type: string
            subscriptionStatus:
              type: string
            startDate:
              type: ["string", "null"]
            expiryDate:
              type: ["string", "null"]
            lastPaymentReference:
              type: ["string", "null"]
            subscriptionNotes:
              type: ["string", "null"]
          required: [id, planType, subscriptionStatus]
        stats:
          type: object
          properties:
            totalAppointments:
              type: integer
            todayAppointments:
              type: integer
            pendingCount:
              type: integer
            confirmedCount:
              type: integer
            completedCount:
              type: integer
            todayOnlineAppointments:
              type: integer
            todayManualAppointments:
              type: integer
            todayBlockedSlots:
              type: integer
            todayCompletedAppointments:
              type: integer
            todayCancelledAppointments:
              type: integer
          required: [totalAppointments, todayAppointments, pendingCount, confirmedCount, completedCount, todayOnlineAppointments, todayManualAppointments, todayBlockedSlots, todayCompletedAppointments, todayCancelledAppointments]
        recentAppointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        settings:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            upiId:
              type: string
            upiQrCodeUrl:
              type: ["string", "null"]
            monthlyPrice:
              type: string
            quarterlyPrice:
              type: string
            yearlyPrice:
              type: string
            supportContact:
              type: string
            supportWhatsapp:
              type: string
          required: [id, upiId, monthlyPrice, quarterlyPrice, yearlyPrice, supportContact, supportWhatsapp]
      required: [clinic, stats, recentAppointments, settings]
...